locator: allocate tablets in specific racks when using rf-rack-valid keyspaces
To avoid cross-rack migrations, we want to use an unchanging set of racks for
replication of each table. To avoid cross-rack view updates, we want this set
of racks to be the same for all tables in a keyspace.

This behavior is enabled by the RF-rack-valid-keyspaces feature. Currently,
it only works when the table was created using all rack in a datacenter.
Instead, we want to be able to enforce this behavior whenever RF is no greater
than the number of rack in a datacenter.
In this patch we prepare the tablet allocation logic to contain the entire
keyspace in RF racks, with one replica of each tablet in each rack.

The racks are selected randomly when creating the first table in the keyspace
(possibly again if all previous tables in the keyspace have been removed).
When we increase RF, all new tablet replicas are allocated in the same rack.
When we decrease RF, we also remove all tablet replicas from a single rack.
When we create a new table in the keyspace and another table already exists there,
the racks for the new table are selected to be the same as the racks used by the
existing table.

To check whether the newly added tablet allocation logic works, a new
unit test is added to boost/network_topology_strategy_test.