tablets: allow creating/altering RF-rack-valid keyspaces with RF<#racks
In this patch we remove the restrictions imposed in former patches regarding
RF-rack-valid keyspaces with RF lower than the number of racks. After the
previous commit, the tablet allocation logic is ready to create/alter
keyspaces as long RF is no greater than the number of racks while using
only RF racks for replication.

The removed restrictions are:
* creating a keyspace with RF < rack count
* lowering the RF of a keyspace to a value greater than 1
* increasing the RF of a keyspace to a value lower than rack count
* creating a table in a keyspace after adding a rack without increasing
  the RF to match rack count

This patch also changes the behavior for datacenters where RF=1. Previously,
when RF was 1, the keyspace was treated as RF-rack-valid, even though its
tablets could migrate between racks. After this patch, the RF=1 case is
treated the same way as other RF values.

With these changes, the RF-rack-valid-keyspaces feature is redefined: a
RF-rack-valid keyspace is now a keyspace that has replicas in RF racks
in each DC, with one replica of each tablet on each rack.

The tests for confirming the previous limitations are removed and the
tests for confirming allowed behavior are extended to match the new
capabilities.

Fixes: https://github.com/scylladb/scylladb/issues/23737