db: service: add new columns to system.topology
Adds two column to system.topology table.

scheduled_rf_change_requests column keeps a list of ids of rf change
request, that wait to be executed. The requests are processed in FIFO
order. The request id is removed from the list after it is done.

ongoing_rf_change_data keeps a data of currently processed rf change
request, namely:
- ks_name;
- new_ks_props - the properties of keyspace to be applied;
- added_racks_for_datacenter - a map of dc to rack list; contains racks
to which a new replica was added (or is being added);
- removed_racks_for_datacenter - analogically as above;
- rollback - a flag that determines whether the operation should be aborted.

During the rollback, the rolled back racks are deleted from added_racks_for_datacenter
and added to removed_racks_for_datacenter (replica removal cannot be rolled back).

In this commit no data is kept in the new columns.