db: service: add new columns to system.topology
Adds two columns to system.topology table.

scheduled_rf_change_requests column keeps a list of ids of rf change
request, that wait to be executed. The requests are processed in FIFO
order. The request id is removed from the list after the request is done.

ongoing_rf_change_data keeps the data of currently processed rf change
request, namely:
- ks_name;
- new_ks_props - the properties of keyspace to be applied;
- added_dc_racks - a list of dcs and racks to which a new
replica was added (or is being added);
- removed_dc_racks - analogically as above;
- rollback - a flag that determines whether the operation should be aborted.

During the rollback, the rolled back racks are deleted from added_dc_racks
and added to removed_dc_racks (replica removal cannot be rolled back).

In this commit no data is kept in the new columns.